Revert "fix(core): 样例题目源目录与题号对齐（A+B=1001，星港舱门=1003）"
This reverts commit 01ba766ea88c922f23ca030d1edc24f1218cc922.
