+++ noj-core/data/problems-src/1001/problem.json
@@ -1,21 +1,21 @@
 {
   "format_version": 1,
   "number": 1001,
-  "title": "1001 T0-LMCC：A+B Problem",
+  "title": "1001 T0-LMCC：星港舱门报码归一化",
   "difficulty": "easy",
   "type": "P",
-  "categories": ["算法"],
+  "categories": ["LMCC 样例题", "算法"],
   "runtime_config": {
     "evaluator": {
       "image": "noj-evaluator-python",
-      "time_limit_ms": 1000,
-      "memory_limit_mb": 256
+      "time_limit_ms": 5000,
+      "memory_limit_mb": 512
     },
     "solution": {
       "image": "noj-solution-python",
       "entry": "submission_sample.py",
       "call_timeout_ms": 5000,
-      "memory_limit_mb": 256
+      "memory_limit_mb": 512
     }
   }
 }
