+++ noj-core/data/problems-src/1001/statement.md
@@ -1,33 +1,113 @@
-# 1001 T0-LMCC：A+B Problem
+# 1001 T0-LMCC：星港舱门报码归一化
+
+## 题目背景
+
+星港是一座用于向太空输送物资的集散港。运行区有一整套连接主压差环、减压环廊、货运气闸和外作业桥的标准舱门网络。任何一扇舱门报码不清，都会影响人员隔离、货箱转运和局部舱段的气压切换。
+
+一线报码通常来自巡检蜂群、拖运司机和夜班值守员，短、急、乱，但主控台需要的是稳定、可解析、可落库的标准
+JSON。你的任务是通过 Prompt 约束，让模型把这些自然语言报码整理成统一结构。
 
 ## 题目描述
 
-给定两个整数 $a$ 和 $b$，计算它们的和。
+对于每条输入报码，模型最终必须输出一个 JSON，包含两个字段：
+
+1. `gate_id`：标准化舱门编号
+2. `status`：标准化舱门状态
+
+`gate_id` 的格式固定为 `X-YY`：
+
+- `X` 为区域字母，只能是 `E / W / N / S / I / O`
+- `YY` 为两位数字编号（`01`–`12`）
+
+区域映射规则如下：
+
+- `E`：东环 / 东区 / 东侧
+- `W`：西环 / 西区 / 西侧
+- `N`：北环 / 北区 / 北侧
+- `S`：南环 / 南区 / 南侧
+- `I`：主环 / 内环 / 内侧
+- `O`：外环 / 外侧
+
+编号规则如下：
+
+1. 需要从文本中识别"几号门"。
+2. 阿拉伯数字直接转为两位数字，例如 `2号门` -> `02`。
+3. 中文数字也要识别，例如 `七号门` -> `07`，`十号门` -> `10`。
+
+`status` 只能是以下三种之一：
+
+- `open`
+- `closed`
+- `fault`
+
+状态按以下优先级判定：
+
+1. 若文本中出现故障语义，如"故障""打不开""拉不开""失灵""卡住""异常""坏了"，则判为
+   `fault`。
+2. 否则，若出现关闭语义，如"关闭""封闭""锁住""关着""暂停通行"，则判为 `closed`。
+3. 否则，若出现开启语义，如"开启""打开""恢复通行""放行""通了""重新打开"，则判为
+   `open`。
+
+也就是说：`fault > closed > open`。
 
 ## 输入格式
 
-一行，两个整数 $a$ 和 $b$，空格分隔。
+输入是一个描述舱门状态的纯文本字符串 `report`（无需解析 JSON），例如：
 
-## 输出格式
+```
+东环七号门现在打不开，报个故障
+```
 
-一行，一个整数，即 $a + b$ 的值。
+## 输出要求
+
+输出标准化的 JSON 字符串，恰含以下两个字段：
+
+- `gate_id` (`string`) — 标准化舱门编号，格式 `X-YY`（区域字母 + 两位编号）
+- `status` (`string`) — 标准化舱门状态（`open` / `closed` / `fault`）
+
+## 评分标准
+
+总分 `10` 分：
+
+- 内容正确 `8` 分：在可见集与隐藏集上联合按字段级准确率计分
+- 格式正确 `2` 分：在可见集与隐藏集上联合按格式命中率计分
+
+格式命中要求：JSON 恰好包含 `gate_id` / `status` 两个键，且 `gate_id` 符合
+`X-YY` 格式（编号 `01`–`12`）、`status` 属于 `open / closed / fault`。
 
 ## 示例
+
+### 示例 1
 
 **输入**
 
 ```
-1 2
+外环七号门完全拉不开
 ```
 
 **输出**
 
-```
-3
+```json
+{ "gate_id": "O-07", "status": "fault" }
 ```
 
-## 限制
+> 注意：输出的 JSON 键名使用 `snake_case`。
 
-- $-10^9 \leq a, b \leq 10^9$
-- 时间限制：$1000\text{ms}$
-- 内存限制：$256\text{MB}$
+### 示例 2
+
+**输入**
+
+```
+主环8号门已经放行
+```
+
+**输出**
+
+```json
+{ "gate_id": "I-08", "status": "open" }
+```
+
+## 参考实现
+
+参考实现见 `submission_sample.py`：在 `solve(report: str) -> str`
+中完成归一化逻辑， 返回标准化 JSON 字符串。
