+++ noj-core/data/problems-src/1003/visible.jsonl
@@ -1,8 +1,10 @@
-{"id":"t0_v001","text":"东环七号门现在打不开，报个故障","expected_gate_id":"E-07","expected_status":"fault"}
-{"id":"t0_v002","text":"请把西区2号门关闭，先别放人","expected_gate_id":"W-02","expected_status":"closed"}
-{"id":"t0_v003","text":"南侧九号门已经恢复通行","expected_gate_id":"S-09","expected_status":"open"}
-{"id":"t0_v004","text":"主环三号门卡住了","expected_gate_id":"I-03","expected_status":"fault"}
-{"id":"t0_v005","text":"北区11号门暂时封闭","expected_gate_id":"N-11","expected_status":"closed"}
-{"id":"t0_v006","text":"外环一号门已开启","expected_gate_id":"O-01","expected_status":"open"}
-{"id":"t0_v007","text":"内侧六号门失灵，完全拉不开","expected_gate_id":"I-06","expected_status":"fault"}
-{"id":"t0_v008","text":"东区12号门重新打开","expected_gate_id":"E-12","expected_status":"open"}
+{"id":"v001","input":"0 0\n","expected":0}
+{"id":"v002","input":"1 2\n","expected":3}
+{"id":"v003","input":"-1 1\n","expected":0}
+{"id":"v004","input":"100 200\n","expected":300}
+{"id":"v005","input":"-5 -7\n","expected":-12}
+{"id":"v006","input":"0 100\n","expected":100}
+{"id":"v007","input":"999 1\n","expected":1000}
+{"id":"v008","input":"-1000 500\n","expected":-500}
+{"id":"v009","input":"12345 67890\n","expected":80235}
+{"id":"v010","input":"7 -3\n","expected":4}
